requirements.txt
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
coverage==7.6.12 ; python_version >= "3.10" and python_version < "4.0"
cryptography==44.0.2 ; python_version >= "3.10" and python_version < "4.0"
deprecated==1.2.18 ; python_version >= "3.10" and python_version < "4.0"
dollar-ref==0.1.3 ; python_version >= "3.10" and python_version < "4.0"
events==0.5 ; python_version >= "3.10" and python_version < "4.0"
gitdb==4.0.12 ; python_version >= "3.10" and python_version < "4.0"
gitpython==3.1.44 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.10" and python_version < "4.0"
loguru==0.7.3 ; python_version >= "3.10" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.10" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "4.0"
opensearch-dsl==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
opensearch-py==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0"
pygithub==1.59.1 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.19.1 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
pynacl==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
rich==13.9.4 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
smmap==5.0.2 ; python_version >= "3.10" and python_version < "4.0"
termcolor==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version == "3.10"
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
win32-setctime==1.2.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
wrapt==1.17.2 ; python_version >= "3.10" and python_version < "4.0"
